Fix DMTCP to also work on i386
- Fixes bug for i386, introduced in rev. 2480
- Warning: This is a major change
  + Brought back mtcp_check_vdso.c to handle randomized vdso
  +  This is enabled for all architectures:  maybe only needed for i386,
       but that's because the other architectures do not cause a collision
       by randomizing vdso into a place where we allocated libraries.

+++ trunk/src/mtcp/tlsinfo.c
@@ -470,7 +470,7 @@ void TLSInfo_VerifyPidTid(pid_t pid, pid_t tid)
 
   if ((tls_pid != pid) || (tls_tid != tid)) {
     PRINTF("ERROR: getpid(%d), tls pid(%d), and tls tid(%d) must all match\n",
-           mtcp_sys_getpid(), tls_pid, tls_tid);
+           (long)mtcp_sys_getpid(), tls_pid, tls_tid);
     _exit(0);
   }
 }
